move component ticks and review comments
