+++ crates/bevy_ecs/src/change_detection.rs
@@ -1,4 +1,4 @@
-use crate::component::{Component, ComponentTicks};
+use crate::component::Component;
 use bevy_reflect::Reflect;
 use std::ops::{Deref, DerefMut};
 
@@ -25,8 +25,8 @@ pub trait DetectChanges {
     /// system.
     fn is_added(&self) -> bool;
 
-    /// Returns true if (and only if) this value been changed since the last execution of this
-    /// system.
+    /// Returns true if (and only if) this value has been mutably accesses since the last execution
+    /// of this system.
     fn is_changed(&self) -> bool;
 }
 
@@ -131,6 +131,71 @@ macro_rules! impl_debug {
         }
 
     };
+}
+
+#[derive(Copy, Clone, Debug)]
+pub struct ComponentTicks {
+    pub(crate) added: u32,
+    pub(crate) changed: u32,
+}
+
+impl ComponentTicks {
+    #[inline]
+    pub fn is_added(&self, last_change_tick: u32, change_tick: u32) -> bool {
+        // The comparison is relative to `change_tick` so that we can detect changes over the whole
+        // `u32` range. Comparing directly the ticks would limit to half that due to overflow
+        // handling.
+        let component_delta = change_tick.wrapping_sub(self.added);
+        let system_delta = change_tick.wrapping_sub(last_change_tick);
+
+        component_delta < system_delta
+    }
+
+    #[inline]
+    pub fn is_changed(&self, last_change_tick: u32, change_tick: u32) -> bool {
+        let component_delta = change_tick.wrapping_sub(self.changed);
+        let system_delta = change_tick.wrapping_sub(last_change_tick);
+
+        component_delta < system_delta
+    }
+
+    pub(crate) fn new(change_tick: u32) -> Self {
+        Self {
+            added: change_tick,
+            changed: change_tick,
+        }
+    }
+
+    pub(crate) fn check_ticks(&mut self, change_tick: u32) {
+        check_tick(&mut self.added, change_tick);
+        check_tick(&mut self.changed, change_tick);
+    }
+
+    /// Manually sets the change tick.
+    /// Usually, this is done automatically via the [`DerefMut`](std::ops::DerefMut) implementation
+    /// on [`Mut`](crate::world::Mut) or [`ResMut`](crate::system::ResMut) etc.
+    ///
+    /// # Example
+    /// ```rust,no_run
+    /// # use bevy_ecs::{world::World, component::ComponentTicks};
+    /// let world: World = unimplemented!();
+    /// let component_ticks: ComponentTicks = unimplemented!();
+    ///
+    /// component_ticks.set_changed(world.read_change_tick());
+    /// ```
+    #[inline]
+    pub fn set_changed(&mut self, change_tick: u32) {
+        self.changed = change_tick;
+    }
+}
+
+fn check_tick(last_change_tick: &mut u32, change_tick: u32) {
+    let tick_delta = change_tick.wrapping_sub(*last_change_tick);
+    const MAX_DELTA: u32 = (u32::MAX / 4) * 3;
+    // Clamp to max delta
+    if tick_delta > MAX_DELTA {
+        *last_change_tick = change_tick.wrapping_sub(MAX_DELTA);
+    }
 }
 
 pub(crate) struct Ticks {
